server/requirements-fastapi.txt
# FastAPI Server Requirements for BloomWatch Kenya

# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0

# Existing backend dependencies
earthengine-api==0.1.418
rasterio==1.3.10
numpy==1.26.4
scipy==1.13.1
xarray==2024.6.0
geopandas==0.14.0
plotly==5.22.0
folium==0.17.0
pytest==8.2.2
africastalking==1.2.6
pymongo==4.6.1
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
pandas==2.1.4

# AI & ML
openai==1.54.3
scikit-learn==1.3.2

# Email
sendgrid==6.11.0

# Additional utilities
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4


